extract run_eagle_verify into eagle_worker_common
